+++ frontend/src/pages/ChatPage/ChatInput.styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@emotion/css';
+
+export const chatInputStyle = css({
+  height: 60,
+  padding: '0 20px',
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  backgroundColor: 'white', // TODO: 상수화
+  borderTop: '1px solid black', // TODO: 상수화
+});
