+++ frontend/src/pages/ChatPage/ChatInput.tsx
@@ -1,0 +1,5 @@
+import { chatInputStyle } from './ChatInput.styles';
+
+export const ChatInput = () => {
+  return <div className={chatInputStyle}>입력칸</div>;
+};
