+++ frontend/src/pages/ChatPage/ChatPage.styles.ts
@@ -27,14 +27,3 @@ export const chatPageListWrapperStyle = css({
   overflowX: 'hidden',
   overflowY: 'scroll',
 });
-
-export const chatInputStyle = css({
-  height: 60,
-  padding: '0 20px',
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  backgroundColor: 'white', // TODO: 상수화
-  borderTop: '1px solid black', // TODO: 상수화
-});
